requirements.txt
asgiref==3.5.2
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.1.0
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.3.0
pytz==2022.7
sqlparse==0.4.3
zipp==3.11.0
